#1 - Map 로직 UI에서 분리
