+++ app/src/main/java/com/demo/desh/MainActivity.kt
@@ -1,11 +1,8 @@
 package com.demo.desh
 
-import android.app.Activity
 import android.content.Context
-import android.graphics.Color
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.compose.setContent
 import androidx.compose.material.Scaffold
 import androidx.compose.material3.Surface
@@ -14,22 +11,18 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.demo.desh.api.RetrofitClient
 import com.demo.desh.model.User
-import com.demo.desh.nav.BottomNavigationBar
-import com.demo.desh.nav.MainBottomBarNav
+import com.demo.desh.util.BottomNavigationBar
+import com.demo.desh.util.MainBottomBarNav
 import com.demo.desh.ui.screens.MainScreen
+import com.demo.desh.ui.screens.MapScreen
 import com.demo.desh.ui.screens.ProfileScreen
 import com.demo.desh.ui.screens.SettingsScreen
 import com.demo.desh.ui.theme.DeshprojectfeTheme
+import com.demo.desh.util.MapCreator
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
-import net.daum.mf.map.api.CameraUpdateFactory
-import net.daum.mf.map.api.MapCircle
-import net.daum.mf.map.api.MapPoint
-import net.daum.mf.map.api.MapPointBounds
 import net.daum.mf.map.api.MapView
 
 class MainActivity : AppCompatActivity() {
@@ -39,8 +32,7 @@ class MainActivity : AppCompatActivity() {
         val user = intent.getSerializableExtra("user") as User
 
         CoroutineScope(Dispatchers.IO).launch {
-            val circles = getMapCircles()
-            val mapView = getMapView(circles)
+            val mapView = MapCreator.getMapView()
 
             runOnUiThread {
                 setContent {
@@ -58,20 +50,28 @@ class MainActivity : AppCompatActivity() {
 @Composable
 fun App(user: User, mapView: (context: Context) -> MapView) {
     val navController = rememberNavController()
+    val onMapButtonClick = { navController.navigate(MainBottomBarNav.Map.route) {
+        restoreState = true
+    } }
 
     Scaffold(
         content = { it
-            MainNavigationHost(navController = navController, user = user, mapView = mapView)
+            MainNavigationHost(navController = navController, user = user, mapView = mapView, onMapButtonClick = onMapButtonClick)
         },
         bottomBar = { BottomNavigationBar(navController = navController) }
     )
 }
 
 @Composable
-fun MainNavigationHost(navController: NavHostController, user: User, mapView: (context: Context) -> MapView) {
+fun MainNavigationHost(
+    navController: NavHostController,
+    user: User,
+    mapView: (context: Context) -> MapView,
+    onMapButtonClick: () -> Unit
+) {
     NavHost(navController = navController, startDestination = MainBottomBarNav.Home.route) {
         composable(route = MainBottomBarNav.Home.route) {
-            MainScreen(mapView)
+            MainScreen(user, onMapButtonClick)
         }
 
         composable(route = MainBottomBarNav.Profile.route) {
@@ -81,55 +81,9 @@ fun MainNavigationHost(navController: NavHostController, user: User, mapView: (c
         composable(route = MainBottomBarNav.Settings.route) {
             SettingsScreen()
         }
-    }
-}
 
-private fun getMapView(circles: List<MapCircle>) : (context: Context) -> MapView {
-    Log.e("getMapView()", "${circles.size}")
-
-    return { context: Context ->
-        val mv = MapView(context)
-        val mapPointBoundsArray = mutableListOf<MapPointBounds>()
-
-        circles.forEach { circle ->
-            mv.addCircle(circle)
-            mapPointBoundsArray.add(circle.bound)
+        composable(route = MainBottomBarNav.Map.route) {
+            MapScreen(mapView)
         }
-
-        val mapPointBounds = MapPointBounds(mapPointBoundsArray.toTypedArray())
-        val padding = 50
-
-        mv.moveCamera(CameraUpdateFactory.newMapPointBounds(mapPointBounds, padding))
-        mv
     }
-}
-
-private suspend fun getMapCircles() : List<MapCircle> {
-    val userService = RetrofitClient.userService
-    val circles = mutableListOf<MapCircle>()
-
-    CoroutineScope(Dispatchers.IO).async {
-        val res = userService.getRecommendationAllInfo()
-
-        if (res.isSuccessful) {
-            val body = res.body()!!
-            val size = body.size
-            val list = body.list
-
-            Log.d("전체 추천 DTO 응답 성공", "size = $size, list = $list")
-
-            list.forEach { res ->
-                val circle = MapCircle(
-                    MapPoint.mapPointWithGeoCoord(res.lat, res.lng),
-                    1000,
-                    Color.argb(255, 255, 255, 0),
-                    Color.argb(128, 128, 128, 128)
-                )
-
-                circles.add(circle)
-            }
-        }
-    }.await()
-
-    return circles
 }
